Refactoring Jitsi cowebsites
Refactoring the Jitsi cowebsites to attempt to avoid race conditions where 2 Jitsis are loaded at once.
Merging JitsiFactory with JitsiCoWebsite. The code is cleaner now and easier to understand.

Also, introducing a new "randomDelay" function that randomly introduces delay in the application when localStorage is set with

```
window.localStorage.setItem("randomDelay", "true")
```

Introducing random delay in callbacks can help pinpoint race conditions.
